1.20 updates and other additions.
Changes for 1.20, as well as additions for log support, watchlists, Wikia compatibility, and various other things I forgot about.

Includes stuff by Jack Phoenix, Skizzerz, Isarra, and possibly one or two others. May or may not be backwards compatible, but it seems to work.
While standards suggest moving the authplugin class to a separate file, to do so at this stage will only result in making the diffs unusable. This can be done later in a separate commit.

Change-Id: Icd85debb59fe8ee9c07b9a821bb77cd04c784007