Make testUserExists() also validate username.
Some wikis seem to display bogus results for list=allusers in the API
even if the aufrom and auto params should make it show 0 results. As
such, validate the username returned in addition to the number of users
to ensure that the user we wanted was actually found or not.

Change-Id: I8526f590e7f64a16be1c5eda22dbd3ce852066f2