Reattribute edits to imported users.
Edits and log entries which previously had user id 0 but the username of
the imported user would stay not-reattributed, which meant it wouldn't
show up in Special:Contributions. Now we (optionally) attribute edits as
well. A maintenance script is provided which can reattribute all edits,
and does not require this extension to be installed to run.

Additionally, a new config option dictates whether or not to import the
watchlist.

Change-Id: Ifd26850a9fdf36deaf3c4408375c1dbd9f80eb86