Ensure creds are always set in primary provider
We need to explicitly set the user's pw in the primary provider on
account creation, as AuthManager makes no assumptions about it, and as
such users were getting set up with null credentials (preventing them
from logging in).

Now that this is resolved, re-enabled the secondary screen to change
password as default.

Change-Id: Icf549f241181976af2398051542f0140b0d62f63