Make edit attribution actor-aware
Edit reattribution jobs got more intelligent as well, and will only
create jobs proportionate to the number of edits that need to be
reattributed rather than the number of revisions on the wiki; this
should greatly speed up imports on large wikis and prevent job queue
bloat.

The edit reattribution now updates the log_search table as well, whereas
that was previously missed. It is recommended to run the
reattributeEdits.php maintenance script after updating from a previous
version to properly convert this table as well.

MediaWikiAuth is compatible with MediaWiki 1.31+ and requires PHP 7.1+.
Its version has been increased to 1.3.0.

Bug: T228815
Change-Id: I679f0a160da282a677a64d35a5c7707e69c41eaa