Remove using User ::addGroup, ::getOptions, ::setOption
User ::addGroup, ::getOptions, ::setOption methods will be hard deprecated.
Further to the the bug I needed to remove only User::getOptions.
But in the near future all usages of soft deprecated methods of User
class in ExternalWikiPrimaryAuthenticationProvider::autoCreatedAccount
will be replaced with UserGroupManager and UserOptionsManager,so I also
replaced other methods of User class with it. In order to achieve it,
the MW version was changed to 1.35+

Bug: T275602
Change-Id: I423b3638257087e19109ffbbd5a90a6ba5d47a86