Import data in stub user mode, add config to control preferences import
When operating in stub user mode
($wgMediaWikiAuthDisableAccountCreation = true), the user's preferences,
watchlist, and user groups can still be imported from the remote wiki
when they initially log in. Edit reattribution still does not happen in
stub user mode, as it is expected that operation has already been
performed by the system administrator during the process of creating the
stub users.

Two new configuration variables were introduced to control user
preference import:

* $wgMediaWikiAuthImportOptions (default ['*']) is an array of user
  preferences to import from the remote wiki. If '*' is present in the
  array, all options will be imported. If the array is empty, no options
  will be imported. Otherwise, the explicitly listed options will be
  imported.
* $wgMediaWikiAuthSkipOptions (default []) is an array of user
  preferences to skip over when importing options from the remote wiki.
  A preference listed here will not be imported, even if the preference
  is also listed in $wgMediaWikiAuthImportOptions. This is useful if you
  wish to import everything except for some particular list of options.

Together, these two configuration variables allow fine-grained control
over which user preferences make it over to the local wiki from the
remote one. As before, if a preference is not recognized by the local
wiki (due to a difference in MediaWiki version, or a difference in
installed extensions, for example), it will not be imported regardless.
Watchlist tokens are additionally never imported.